% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean Input Text}
\usage{
clean_text(inputText, replaceSymbol = FALSE, replaceNumber = FALSE)
}
\arguments{
\item{inputText}{A character string or vector of character strings}

\item{replaceSymbol}{If TRUE, symbols are replaced with their equivalent (e.g. "@" becomes "at"). Defaults to FALSE.}

\item{replaceNumber}{If TRUE, numbers are replaced with their equivalent (e.g. "20" becomes "twenty", "3rd" becomes "third"). Defaults to FALSE.}
}
\value{
A character string (or vector of character strings) with cleaned text.
}
\description{
This function cleans text by: \cr
-Setting all text to lowercase \cr
-Removing non-ASCII characters \cr
-Expanding contractions ("don't" --> "do not") \cr
-Removing punctuation \cr
-Removing symbols (if replaceSymbol is FALSE) \cr
-Removing numbers (if replaceNumber is FALSE)
}
\examples{
myString = "He gave his last $10 to Sally's sister because she's nice."

cleanText = clean_text(myString)
# "he gave his last to sally sister because she is nice"

cleanText = clean_text(myString, replaceNumber = TRUE)
# "he gave his last ten to sally sister because she is nice"

cleanText = clean_text(myString, replaceSymbol = TRUE)
# "he gave his last dollar to sally sister because she is nice"
}
