% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_model.R
\docType{class}
\name{compModel-class}
\alias{compModel-class}
\alias{compModel}
\title{compModel Class}
\description{
compModel Class
}
\section{Slots}{

\describe{
\item{\code{call}}{The function called to generate this model, with all arguments specified by the user}

\item{\code{data_predictor}}{The predictor variable input to create the model}

\item{\code{data_outcome}}{The outcome variable input to create the model}

\item{\code{type}}{Model type, "binary" or "continuous"}

\item{\code{predictor}}{The name of the column in the test dataframe containing the data_predictor}

\item{\code{outcome}}{The name of the column in the test dataframe containing the data_outcome}

\item{\code{y}}{The dependent (outcome) variable}

\item{\code{glm}}{The general linear model created}

\item{\code{predicted_y}}{The predicted outcomes based on the model and original predictor data}

\item{\code{predicted_probabilities}}{(If binary) The predicted probabilities of the outcomes based on the model and original predictor data}

\item{\code{roc}}{(If binary) The ROC calculated using the predicted_y}

\item{\code{roc_ci}}{(If binary) The boostrapped confidence interval calculated for the ROC}

\item{\code{corr}}{(If continuous) The correlation using the predicted_y}

\item{\code{level0}}{The bottom/first level of a binary variable, or the lowest value of a continuous variable}

\item{\code{level1}}{The top/second level of a binary variable, or the highest value of a continuous variable}
}}

