% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idiosync.R
\name{idiosync_participant_words}
\alias{idiosync_participant_words}
\title{Idiosyncratic Participant Words}
\usage{
idiosync_participant_words(
  inputDataframe,
  mode,
  textColumnName,
  participantColumnName
)
}
\arguments{
\item{inputDataframe}{A dataframe containing a column with text data (character strings) and participant IDs}

\item{mode}{This defines the mode of operation. Options include "output", "remove", or "both". See Details below.}

\item{textColumnName}{A string consisting of the name of the column in \code{inputDataframe} which contains text data}

\item{participantColumnName}{A string consisting of the name of the column in \code{inputDataframe} which contains participant IDs}
}
\value{
A dataframe (\code{mode="output"}), a character string or vector of character strings (\code{mode="remove"}), or a two-object list containing both results (\code{mode="both"})
}
\description{
This function identifies participants' words that are idiosyncratic (i.e. used multiple times by a single participant, and never by any other participant). It can also be used to remove these words.
}
\details{
This function has three modes:
In the "output" mode, a dataframe is produced with three columns: the participant who produced the idiosyncratic words, the words, and how frequently they are used by the participant.
In the "remove" mode, a character string (or vector of character strings) is produced, where all of the idiosyncratic words are removed.
In the "both" mode, both of the above results will be produced (i.e. a list containing a dataframe of idiosyncratic words, as well as the text with those words removed)
}
\examples{
myStrings = c("Last week while I was walking in the park, I saw a firetruck go by. It was red.",
              "My dog loves to go on walks in the park every day of the week.",
              "Where I live, it snows all winter long. It's so cold outside.",
              "My kids love to play in the snow. They love to collect snow to build snowmen.",
              "When I was younger, I used to visit my grandmother every week.",
              "In the summertime, we would get together with my grandmother to bake cookies.")
mydataframe = data.frame(text=myStrings, participant=c(1,1,2,2,3,3), stringsAsFactors = FALSE)
idiosync_output = idiosync_participant_words(mydataframe, "output", "text", "participant")
idiosync_output
# participant     feature       frequency
# 1                park          2
# 1                go            2
# 2                love          2
# 2                snow          2
# 3                grandmother   2

idiosync_removed = idiosync_participant_words(mydataframe, "remove", "text", "participant")
idiosync_removed
# "Last week while I was walking in the, I saw a firetruck by. It was red."
# "My dog loves to on walks in the every day of the week."
# "Where I live, it snows all winter long. It's so cold outside."
# "My kids to play in the. They to collect to build snowmen."
# "When I was younger, I used to visit my every week."
# "In the summertime, we would get together with my to bake cookies."


}
\seealso{
\code{\link{idiosync_response_words}}
}
