% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_contractions_addl}
\alias{key_contractions_addl}
\title{Contraction Data}
\format{
A data frame with 140 rows and 2 variables:
\describe{
  \item{contraction}{the original contraction (e.g. "don't" or "dont")}
  \item{expanded}{the expanded form (e.g. "do not")}
}
}
\source{
\url{https://github.com/trinker/lexicon/blob/master/data/key_contractions.rda}
}
\usage{
key_contractions_addl
}
\description{
A dataset containing versions of contractions to be used with \code{textclean} that don't contain apostrophes (e.g. it will still convert "dont" to "do not").
}
\keyword{datasets}
