% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_model.R
\docType{class}
\name{langModel-class}
\alias{langModel-class}
\alias{langModel}
\title{langModel Class}
\description{
langModel Class
}
\section{Slots}{

\describe{
\item{\code{data_text}}{The text input to create the corpus/model}

\item{\code{data_outcome}}{The outcome variabe input to create the model}

\item{\code{type}}{Model type, "binary" or "continuous"}

\item{\code{text}}{The name of the column in the original_dataframe containing the data_text}

\item{\code{outcome}}{The name of the column in the original_dataframe containing the data_outcome}

\item{\code{tokens}}{The list of tokens in the language corpus}

\item{\code{x}}{The document-frequency matrix}

\item{\code{y}}{The dependent (outcome) variable}

\item{\code{cv}}{The final model}

\item{\code{lambda}}{The lambda value used}

\item{\code{level0}}{The bottom/first level of a binary variable, or the lowest value of a continuous variable}

\item{\code{level1}}{The top/second level of a binary variable, or the highest value of a continuous variable}

\item{\code{cat0raw}}{The predictors (word engrams) predicting the level0 outcome, with their model weights}

\item{\code{cat1raw}}{The predictors (word engrams) predicting the level1 outcome, with their model weights}

\item{\code{p_value}}{The p-value estimated via permutation test}
}}

