% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lemma_data}
\alias{lemma_data}
\title{Lemma Data}
\format{
A data frame with 47,366 rows and 2 variables:
\describe{
  \item{inflected_form}{the original, inflected word (e.g. 'dogs' or 'walked')}
  \item{lemma}{the base form of the word (e.g. 'dog' or 'walk')}
}
}
\source{
\url{https://github.com/tm4ss/tm4ss.github.io/blob/master/resources/baseform_en.tsv}
}
\usage{
lemma_data
}
\description{
A dataset containing inflected and base forms of text, used for the \code{lemmatize} function.
}
\keyword{datasets}
