% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmatize.R
\name{lemmatize}
\alias{lemmatize}
\title{Lemmatize Text}
\usage{
lemmatize(inputText, treetaggerDirectory, progressBar = TRUE)
}
\arguments{
\item{inputText}{A character string or vector of character strings}

\item{treetaggerDirectory}{the filepath to the location of your installation of the \code{treetagger} library (See Details below)}

\item{progressBar}{Show a progress bar. Defaults to TRUE.}
}
\value{
A dataframe with lemmatized text, as well as columns with information about parts of speech
}
\description{
This function performs lemmatization on input text by reducing words to their base units.
}
\details{
This function is essentially a wrapper for the \code{\link[koRpus]{treetag}} function from the \code{[koRpus]} package.
In turn, koRpus implements the TreeTagger software package (available here: \url{https://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}).
The software must be downloaded and installed on your local computer in order to use the \code{lemmatize} function.
Once installed, the \code{treetaggerDirectory} argument should consist of the path where the software was installed.

This function performs "lemmatization," which is one form of reducing words to their most basic units. It is more thorough than "stemming," which only removes suffixes.
E.g. for the words "walked" and "dogs," both lemmatization and stemming would reduce the words to "walk" and "dog."
However, stemming would ignore "ran" and "geese," while lemmatization would properly render these "run" and "goose."
}
\examples{
myStrings = c("I walked in the park with both of my dogs.",
"The largest geese ran very fast.")
\dontrun{
lemmatized_data = lemmatize(myStrings, "~/path/to/TreeTagger")
lemmatized_data$lemma_text}
# "I walk in the park with both of my dog."
# "The large goose run very fast."

}
\seealso{
the \code{\link[koRpus]{treetag}} function from the \code{koRpus} package, as well as the treetagger documentation: \url{https://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}
}
