% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_models.R
\docType{class}
\name{modelAssessment-class}
\alias{modelAssessment-class}
\alias{modelAssessment}
\title{modelAssessment Class}
\description{
modelAssessment Class
}
\section{Slots}{

\describe{
\item{\code{type}}{Model type, "binary" or "continuous"}

\item{\code{model_labels}}{Label names and basic info for each model in the assessment}

\item{\code{auc_polygon_df}}{Dataframe for plotting the AUC polygon}

\item{\code{roc_ci_df}}{Dataframe for plotting the ROC confidence intervals}

\item{\code{roc_curve_df}}{Dataframe for plotting the ROC curves}

\item{\code{auc_ci_labels_df}}{Dataframe for AUC labels}

\item{\code{auc_tests}}{Dataframe of significance tests for AUCs}

\item{\code{cat_data}}{Dataframe of predictors (word engrams) with their model weights}
}}

