% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_review_data1}
\alias{movie_review_data1}
\title{Movie Reviews from IMDB}
\format{
A data frame with 2000 rows and 3 variables:
\describe{
  \item{text}{text of the user's movie review}
  \item{valence}{valence of the user's movie review, \code{Positive} (rating 1-4) or \code{Negative} (rating 7-10)}
  \item{rating}{user's rating of the movie, on a scale of 1-10}
}
}
\source{
\url{http://ai.stanford.edu/~amaas/data/sentiment/}
}
\usage{
movie_review_data1
}
\description{
A dataset containing the text and ratings of 2000 movie reviews from IMDB. 1000 are negative (rating 1-4) and 1000 are positive (rating 7-10).
}
\keyword{datasets}
