% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_plots.R
\name{overview_plots}
\alias{overview_plots}
\title{Overview Plots}
\usage{
overview_plots(inputDataframe, textColumnName, participantColumnName)
}
\arguments{
\item{inputDataframe}{A dataframe containing a column with text data (character strings)}

\item{textColumnName}{A string consisting of the name of the column in \code{inputDataframe} which contains text data}

\item{participantColumnName}{(Optional argument) A string consisting of the name of the column in \code{inputDataframe} which contains participant IDs}
}
\value{
Nothing (this function plots a series of graphs)
}
\description{
This function creates a set of plots showing basic information about a corpus of text data.
}
\details{
If a \code{participantColumnName} is not provided, three graphs will be produced:
-A pie chart with the total number of words in the provided corpus, divided into "Unique" words (those only used once), and "Repeated" words (those used at least twice).
-A density plot with the length of each individual response in the corpus (in words)
-A bar plot of the 25 most common words in the corpus, arranged by frequency

If a \code{participantColumnName} is provided, an additional two graphs will be produced:
-The average number of words per response, plotted by participant (the overall average will be displayed as a vertical line)
-The total number of words produced by each participant, compared to the total number of unique words produced by each participant
}
\examples{
\dontrun{
overview_plots(movie_review_data1, "text")
}

}
