% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_network.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot Individual Cluster}
\usage{
plot_cluster(network_input, cluster_number)
}
\arguments{
\item{network_input}{The output of the \code{plot_word_network} function}
}
\value{

}
\description{
Plots individual clusters from the \code{plot_word_network} function. Can be helpful if these need to be plotted separately for any reason.
}
\examples{
\dontrun{
movie_review_data1$cleanText = clean_text(movie_review_data1$text)

# Using language to predict "Positive" vs. "Negative" reviews
movie_model_valence = language_model(movie_review_data1,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText")

node_edge_table = node_edge(movie_model_valence)
network_output = plot_word_network(node_edge_table, clusterType="node")
plot_cluster(network_output[[1]], cluster_number=1)
}
}
