% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictor_words.R
\name{plot_predictor_words}
\alias{plot_predictor_words}
\title{Plot Predictor Words}
\usage{
plot_predictor_words(
  modelAssessment,
  topX,
  colors = c("blue", "orange"),
  plot_titles,
  model_names,
  xaxis_range,
  standard_xaxis = TRUE,
  flip_graphs = FALSE,
  print_individual = TRUE,
  print_summary = TRUE
)
}
\arguments{
\item{modelAssessment}{The output of the \code{\link{assess_models}} function}

\item{topX}{The number of most-predictive words to plot}

\item{colors}{A two-element vector containing the colors of the plotted bars. Defaults to c("blue", "orange")}

\item{plot_titles}{A vector of titles for the plots}

\item{model_names}{A vector of names for the individual models}

\item{xaxis_range}{A maximum value for the x-axis}

\item{standard_xaxis}{If TRUE, the x-axis on all graphs will be the same. If FALSE, it will adjust to fit each individual graph. Defaults to TRUE.}

\item{flip_graphs}{Flips the graphs horizontally. Defaults to FALSE (low-value outcome variable on the left, high-value outcome variable on the right)}

\item{print_individual}{If TRUE, prints an individual graph for each model. Defaults to TRUE.}

\item{print_summary}{If TRUE, prints a summary graph with all models. Defaults to TRUE.}
}
\value{
Nothing (this function plots a series of graphs)
}
\description{
This function plots predictive words from the results of the \code{\link{assess_models}} function
}
\examples{

strong_movie_review_data$cleanText = clean_text(strong_movie_review_data$text)
mild_movie_review_data$cleanText = clean_text(mild_movie_review_data$text)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for strong reviews (ratings of 1 or 10)
movie_model_strong = language_model(strong_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for mild reviews (ratings of 4 or 7)
movie_model_mild = language_model(mild_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Create the model assessment
movie_assessment = assess_models(movie_model_strong, movie_model_mild)

# Analyze ROC curves
plot_predictor_words(movie_assessment)
}
\seealso{
\code{\link{assess_models}}
}
