% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves}
\usage{
plot_roc(
  modelAssessment,
  individual_plot = TRUE,
  combined_plot = TRUE,
  facet_plot = TRUE,
  facet_summary = TRUE,
  colors,
  model_names,
  plot_auc_polygon = TRUE,
  plot_ci = TRUE,
  line_size = 1,
  print_auc = TRUE,
  print_ci = TRUE,
  print_auc_ci_font_size = 4,
  print_auc_ci_x,
  print_auc_ci_y,
  plot_legend = TRUE,
  plot_title,
  facet_n_row = NULL,
  facet_n_col = 2
)
}
\arguments{
\item{modelAssessment}{The output of the \code{\link{assess_models}} function}

\item{individual_plot}{If TRUE, graphs individual ROC curves for each model. Defaults to TRUE.}

\item{combined_plot}{If TRUE, and modelAssessment contains multiple models, graphs a plot with all ROC curves overlapping. Defaults to TRUE.}

\item{facet_plot}{If TRUE, and modelAssessment contains multiple models, graphs a faceted plot with all ROC curves included. Defaults to TRUE.}

\item{facet_summary}{If TRUE, and modelAssessment contains multiple models, the facet_plot will include a plot with all ROC curves overlapping. Defaults to TRUE.}

\item{colors}{A vector of colors to use for each model's ROC curve.}

\item{model_names}{A vector of strings to use as titles/names for each model.}

\item{plot_auc_polygon}{If TRUE, the area below with ROC curve with the lowest AUC will be shaded in. Defaults to TRUE.}

\item{plot_ci}{If TRUE, a confidence band will be plotted around each ROC curve. Defaults to TRUE.}

\item{line_size}{A numeric representing the width of the ROC curve line. Defaults to 1.}

\item{print_auc}{If TRUE, the value of the AUC will be printed on the plot. Defaults to TRUE.}

\item{print_ci}{If TRUE, the range of the confidence interval will be printed on the plot. Defaults to TRUE.}

\item{print_auc_ci_font_size}{The font size for printed values for the AUC and confidence interval. Defaults to 4.}

\item{print_auc_ci_x}{A vector of x (horizontal) positions determining where on the plot the AUC and confidence interval values will be printed.}

\item{print_auc_ci_y}{A vector of y (vertical) positions determining where on the plot the AUC and confidence interval values will be printed.}

\item{plot_legend}{If TRUE, a legend will be printed on all plots.}

\item{plot_title}{The title of the plot}

\item{facet_n_row}{The number of rows used to plot the facet_plot. Defaults to NULL.}

\item{facet_n_col}{The number of columns used to plot the facet_plot. Defaults to 2.}
}
\value{
Nothing (this function plots a series of graphs)
}
\description{
This function plots ROC curves from the results of the \code{\link{assess_models}} function
}
\examples{

\dontrun{
strong_movie_review_data$cleanText = clean_text(strong_movie_review_data$text)
mild_movie_review_data$cleanText = clean_text(mild_movie_review_data$text)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for strong reviews (ratings of 1 or 10)
movie_model_strong = language_model(strong_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for mild reviews (ratings of 4 or 7)
movie_model_mild = language_model(mild_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Create the model assessment
movie_assessment = assess_models(movie_model_strong, movie_model_mild)

# Plot ROC curves
plot_roc(movie_assessment)
}
}
\seealso{
\code{\link{assess_models}}
}
