% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_model.R
\name{summary.compModel}
\alias{summary.compModel}
\title{Summary (compModel)}
\usage{
\method{summary}{compModel}(object, ...)
}
\arguments{
\item{object}{The compModel object to summarize}

\item{...}{Additional arguments}
}
\description{
Summary (compModel)
}
