% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_model.R
\name{summary.langModel}
\alias{summary.langModel}
\title{Summary (langModel)}
\usage{
\method{summary}{langModel}(object, ...)
}
\arguments{
\item{object}{The langModel object to summarize}

\item{...}{Additional arguments}
}
\description{
Summary (langModel)
}
