% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_language_model.R
\name{test_language_model}
\alias{test_language_model}
\title{Test Language Model}
\usage{
test_language_model(
  input,
  outcome,
  text,
  trainedModel,
  ngrams = "1",
  dfmWeightScheme = "count",
  progressBar = TRUE
)
}
\arguments{
\item{input}{A dataframe containing a column with text data (character strings) and an outcome variable (numeric or two-level factor)}

\item{outcome}{A string consisting of the column name for the outcome variable in \code{inputDataframe}}

\item{text}{A string consisting of the column name for the text data in \code{inputDataframe}}

\item{trainedModel}{A trained model created by the \code{\link{language_model}} function}

\item{ngrams}{A string defining the ngrams to serve as predictors in the model. Defaults to "1". For more information, see the \code{okens_ngrams} function in the \code{quanteda} package}

\item{dfmWeightScheme}{A string defining the weight scheme you wish to use for constructing a document-frequency matrix. Default is "count". For more information, see the \code{dfm_weight} function in the \code{quanteda} package}

\item{progressBar}{Show a progress bar. Defaults to TRUE.}
}
\value{
An object of the type "testAssessment"
}
\description{
This function tests a model created by the \code{\link{language_model}} function on a new dataset
}
\details{
This function is effectively a special version of the \code{\link{language_model}} function. Instead of creating a new model, the outputs are based on the results of testing a new, independent dataset using an existing model.
This allows for assessing how well a trained language model generalizes to other inputs - this function allows for comparisons between the models using many of the same functions that can be used with \code{\link{language_model}}.
}
\examples{

\dontrun{
movie_review_data1$cleanText = clean_text(movie_review_data1$text)
movie_review_data2$cleanText = clean_text(movie_review_data2$text)

# Train a model on the \code{movie_review_data1} dataset
# Using language to predict "Positive" vs. "Negative" reviews
movie_model_valence = language_model(movie_review_data1,
                                     outcome = "valence",
                                     outcomeType = "binary",
                                     text = "cleanText")

# Test the model on the \code{movie_review_data2} dataset
movie_model_valence_test = test_language_model(movie_review_data2,
                                    outcome = "valence",
                                    text = "cleanText",
                                    trainedModel = movie_model_valence)
summary(movie_model_valence_test)
}

}
\seealso{
\code{\link{language_model}}
}
