% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_language_model.R
\name{summary.testAssessment}
\alias{summary.testAssessment}
\title{Summary (testAssessment)}
\usage{
\method{summary}{testAssessment}(object, ...)
}
\arguments{
\item{object}{The testAssessment object to summarize}

\item{...}{Additional arguments}
}
\description{
Summary (testAssessment)
}
