% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_roc.R
\name{analyze_roc}
\alias{analyze_roc}
\title{Analyze ROC Curves}
\usage{
analyze_roc(modelAssessment, plot = TRUE, plot_diagonal = FALSE)
}
\arguments{
\item{modelAssessment}{The output of the \code{\link{assess_models}} function}

\item{plot}{If TRUE, plots a matrix displaying the results of all model comparisons. Defaults to TRUE.}

\item{plot_diagonal}{if TRUE, the matrix plot will show repeated (inverted) values on the opposite diagonal. Defaults to FALSE.}
}
\value{
A dataframe with the results of statistical tests conducted on the ROCs for each model pairing
}
\description{
This function analyzes ROC curves from the results of the \code{\link{assess_models}} function
}
\examples{

\dontrun{
strong_movie_review_data$cleanText = clean_text(strong_movie_review_data$text)
mild_movie_review_data$cleanText = clean_text(mild_movie_review_data$text)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for strong reviews (ratings of 1 or 10)
movie_model_strong = language_model(strong_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Using language to predict "Positive" vs. "Negative" reviews
# Only for mild reviews (ratings of 4 or 7)
movie_model_mild = language_model(mild_movie_review_data,
                                     outcomeVariableColumnName = "valence",
                                     outcomeVariableType = "binary",
                                     textColumnName = "cleanText",
                                     progressBar = FALSE)

# Create the model assessment
movie_assessment = assess_models(movie_model_strong, movie_model_mild)

# Analyze ROC curves
auc_tests = analyze_roc(movie_assessment)
}
}
\seealso{
\code{\link{assess_models}}
}
